// types/messagetypes.ts
export enum messageTypes {
  messageText = 'messageText',
  messagePhoto = 'messagePhoto',
  messageDocument = 'messageDocument',
  messageWebPage = 'messageWebPage',
  messageContact = 'messageContact',
  messagePoll = 'messagePoll',
  messageStory = 'messageStory',
  messageAlbum = 'messageAlbum'
}